feat: Allow multiple --as args.
The intent is that this allows values with disjoint sets of "as" options
to be declared and used at the same time. For example:

```toml
env = {dev = "dev", prod = "prod"}
feature = {foo = "feat-1", bar = "feat-2"}
```

You could then use `sauce -a dev -a foo`, whereas previously you'd have
had to execute `sauce` twice while globbing/filtering the sets of keys
being targeted.

Note, that if you were to do `sauce -a dev -a prod` the first value will
be selected.

---
Also, values without a corresponding "as" (for example `sauce -a meow`),
will no longer have their values written to empty.

Also, if an "default" option exists, it will now be selected as a
fallback in all cases where no other value is matched.